+++ www/app/database/migrations/2013_12_09_015637_add_user_id_to_card.php
@@ -15,7 +15,7 @@ class AddUserIdToCard extends Migration {
 		//
         Schema::table('card', function($table)
         {
-            $table->integer('user_id');
+            $table->bigInteger('user_id');
         });
 	}
